==> class_exeRcises/04_installing_packages_Reading_files.Rmd <==
---
title: "04_installing_packages_Reading_files.Rmd"
output: html_document
editor_options: 
  chunk_output_type: console
---


git clone git@github.com:boulderrinnlab/CLASS_2022.git


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(stringsAsFactors = FALSE)
library(GenomicRanges)
library(rtracklayer)
library(tidyverse)
library(ggpubr)
library(Gviz)
source("../../../util/_setup.R")
source("../../../util/intersect_functions.R")
source("../../../util/plotting_functions.R")
```

# Welcome to RStudio-viz! Let's take a little tour.

First we need to learn how to install packages. Let's find out what version of R we are using.

Simply type version in the console.

Now we will use Install.packages() for most packages. This will install any packages in CRAN repository. Let's try one.

```{R install packages}
?install.packages
install.packages("tidyverse")
```

Cool we have our first package installed.

However sometimes we ned to install from Bioconductor and need to use BiocManager::install function

```{R install BiocManager}
if (!requireNamespace("BiocManager", quietly = TRUE))
install.packages("BiocManager", version = '3.10')
```

# Now try installing the packages listed above. 

Some packages like GenomicRanges are from Bioconductor and those will
need to be installed with the BiocManager::install function.

```{R install genomic ranges}
# Here is an example BiocManager::install
BiocManager::install("GenomicRanges")
# try installing all the packages listed above.
# note use either install.packages or BiocManager::install
rtracklayer
tidyverse
```


# ------------------------- R basic data analysis  -------------------------
# All the files needed for this exeRcise can be downloaded here 

```{BASH DOWNLOAD DATA NEEDED FOR LESSON IF NOT IN CLASS}
#  https://www.dropbox.com/sh/76etjxl10v2hcuc/AAAIk9a6aMhw8CvUtKlFnMn6a?dl=0
# wget (url_above) # check your directory.
```

Now let's use an example data file we will produce and how we can use R to do
some quick analyses.

We will addres the following two questions as an example:
# How many peaks does each DBP have? 
# How much of the genome do they cover?

```{r read in peaks per DBP}
# read in csv file that has number of peaks and genome coverage
peaks_per_dbp <- read_csv("/scratch/Shares/rinnclass/CLASS_2022/data/number_of_peaks_per_dbp.csv")
# Now we have an object in the environment that is called 'peaks_per_dbp'
head(peaks_per_dbp)
# This is a tibble or data frame object you can think of like excel sheet for now
# More on Tibble here:
# https://tibble.tidyverse.org/
```

# Using 'table' & 'summary to get information from files
```{R table and summary}
# Let's get more infomration about these peaks with table
table(peaks_per_dbp$dbp)
# Not very useful in this case because there is only one entry for each DBP
summary(peaks_per_dbp$num_peaks)
# Ok this is super helpful - we see the quartiles for each of the categories
```

# Which DBP has the most or least peaks?

```{r DBP max min peaks}
# We can take a shortcut using 'summary' to find out the smallest and largest
summary(peaks_per_dbp$num_peaks)
# we can use a greater or less than function to find the biggest one
peaks_per_dbp$num_peaks > 40000
# We see this produces the T and F for each DBP and we can see 2 DBPs (but we don't know the name :(
```

# Now we will use filter which will allow us to do much more.

```{R filter largets number of peaks}
#To solve this we can use filter and set a variable using <-
max_peaks <- dplyr::filter(peaks_per_dbp, num_peaks > 40000)
# Let's see what else we can do with filter
?filter
#Let's filter on a couple of factors
max_peaks <- dplyr::filter(peaks_per_dbp, num_peaks > 30000, total_peak_length == 133310967)
```


# Let's find the DBP(s) with the fewest peaks (we saw in summary it was 263)
```{r filter smallest number of peaks}
summary(peaks_per_dbp$num_peaks)
# we see that 263 is the min number of peaks.
min_peaks <- filter(peaks_per_dbp, num_peaks < 264)
```

# Let's find the number of peaks for a specific gene say POLR2A
```{r peaks for a specific DBP}
# let's pull out polr2a peaks 
gene_pol2 <- filter(peaks_per_dbp, dbp == "POLR2A")
```


# Now let's make a new column in the file table
```{r making a new column and populating}
# making a column of what
peaks_per_dbp$newcol <- "What?"
```

# Now let's assign a number to each DBP in another new column
```{R new column populating with a function}
peaks_per_dbp$id <- 1:nrow(peaks_per_dbp)
```

# Indexing
```{r indexing rows 1:5 and column 1}
peaks_per_dbp[1:5, 1]
# IMPORTANT : R is always indexes by rows, then columns (comma seperates)
# So above we see take rows one through five and column 1 information.
# More on R indexing https://stats.idre.ucla.edu/r/modules/subsetting-data/
```

# Let's remove a column using indexing
```{R remove a column using indexing}
# let's remove the "what" column we just made above
peaks_per_dbp <- peaks_per_dbp[,-4]
```


# Now let's make a plot of the data. First a histogram of number of peaks for each DBP
```{r}
?hist
hist(peaks_per_dbp$num_peaks, breaks = 100) 
# saving the figure (we will do this over and over again :)
# we will learn ggplot later but we use PDF and turn of null device (plot window) 
pdf("figures/hist_of_peaks_per_dbp.pdf")
hist(peaks_per_dbp$num_peaks, breaks = 100) 
dev.off()
# note always check working directory before saving.
```

# Now let's plot the number of peaks versus peak width.
```{r number of peaks versus genome coverage}
# we will use base R to plot the num_peaks versus genome coverage
pdf("figures/num_peaks_vs_coverage.pdf")
plot(x = peaks_per_dbp$num_peaks, y = peaks_per_dbp$total_peak_length)
dev.off()
```

# Now let's dig into the data file to subset it into the genes we are interested in.
```{r selcting a list of genes from data frame}
# First let's make a list of genes we want. 
# To make a vector of characters we use the c() function
genes <- c("POLR2A","POLR2B", "SUPT5H")
# Now lets make it a searchable index with paste0 (?paste0)
pattern_to_search <- paste0(genes, collapse = "|")
```

# It's worth noting that things can be accomplished in many different ways in R.
# Here we'll get to the same subset of row entries two different ways. 
```{r}
#first let's use grepl
# ?grepl
grepl(pattern_to_search, peaks_per_dbp$dbp)
# table will give us a way to test if we retrieve the three DBPs
table(grepl(pattern_to_search, peaks_per_dbp$dbp))
# Now let's get all the information for these genes by indexing peaks_per_dbp
# So df[,] = all rows/cols, df[1,] = first row all columns etc.
ppdbp_subset <- peaks_per_dbp[grepl(pattern_to_search, peaks_per_dbp$dbp),]
# note we only took the rows for the genes of interest. Then all the cols with (,)
# R syntax is rows,columns.
# df[,] = all rows/cols
# df[1,] = first row all columns
# df[,1] = all rows and column 1
```

# Now let's do the same thing with filter in tidyverse syntax ' %in% ' 
```{R using %in%}
# %in% will find all the data "in" the list being searched "genes"
peaks_per_dbp$dbp %in% genes
logical_vector <- peaks_per_dbp$dbp %in% genes
# Now we will set a variable as the output using filter and then the code we just ran
ppdbp_subset <- dplyr::filter(peaks_per_dbp, dbp %in% genes)
```

# MUTATE !!


# Now let's run functions on the data in a given column with MUTATE (my favorite function)
# head(peaks_per_dbp) -- we can see that the total length is quite long, let's make it in Millions
```{r mutate}
#let's take a look at mutate
?mutate
# Now let's use mutate to divide the total coverage by 1M for every value
peaks_per_dbp <- peaks_per_dbp %>%
  dplyr::mutate(tpl_m = total_peak_length / 1000000)
# syntax is mutate(new_col_name = function(x) in this case divide :)
# instead of making a new col we can set mutate output to old col name
#peaks_per_dbp <- peaks_per_dbp %>% 
 # mutate(total_peak_length = total_peak_length / 1000000) 
# Now let's plot it
hist(peaks_per_dbp$total_peak_length)
```

# Now let's explore the power of mutate syntax with a function.
# We will try it with a more complex function using ifelse logic

```{r mutate if else statement}
peaks_per_dbp <- peaks_per_dbp %>% 
  mutate(genome_coverage_size = ifelse(tpl_m < 1, "small",
                          ifelse(tpl_m < 30, "medium", "large")))
# that is really handy for changing stuff around in a data frame -- 
# note that if else moves forward for greater that 30 and if true then returns large
# In other words if else you only need to conditions to get three values ...
```


*********************
EXCERCISE
*********************

Use what we learned above to plot a historgram of all the proteins that are in the medium sized category for total genome coverage.

Make a table with just the medium coverage DBPs
